Add DTO folder
