+++ internal/model/dto/profile.go
@@ -1,0 +1,5 @@
+package dto
+
+type ProfileUpdate struct {
+  // TODO: Implement
+}
