--- internal/model/role.go
@@ -1,9 +1,0 @@
-package model
-
-import "gorm.io/gorm"
-
-type Role struct {
-	gorm.Model
-	Name  string
-	Users []User `gorm:"many2many:user_roles"`
-}
