+++ internal/model/user.go
@@ -56,41 +56,3 @@ type User struct {
 
 	// TODO: Implement enums Language, Theme and Country (country shouldnt be an enum)
 }
-
-type UserRegister struct {
-	FirstName string `validate:"required,min=3"                  json:"firstname"`
-	LastName  string `validate:"required,min=3"                  json:"lastname"`
-	Username  string `validate:"required,username,min=3,max=32"  json:"username"`
-	Email     string `validate:"required,email"                  json:"email"`
-	Password  string `validate:"required,password,min=12,max=48" json:"password"`
-}
-
-type UserLogin struct {
-	Username string `validate:"omitempty,username,min=3,max=32" json:"username"`
-	Email    string `validate:"omitempty,email"                 json:"email"`
-	Password string `validate:"required,password,min=12,max=48" json:"password"`
-}
-
-type ProfileUpdate struct {
-	FirstName string `validate:"omitempty,min=3"                 json:"firstname"`
-	LastName  string `validate:"omitempty,min=3"                 json:"lastname"`
-	Username  string `validate:"omitempty,username,min=3,max=32" json:"username"`
-	Email     string `validate:"omitempty,email"                 json:"email"`
-	Bio       string `validate:"omitempty"                       json:"bio"`
-	Show      struct {
-		Profile    int `json:"profile"`
-		Image      int `json:"image"`
-		Comments   int `json:"comments"`
-		Favorites  int `json:"favorites"`
-		Projects   int `json:"projects"`
-		Components int `json:"components"`
-		Followers  int `json:"followers"`
-		Following  int `json:"following"`
-		Inventory  int `json:"inventory"`
-		Formations int `json:"formations"`
-	} `json:"show"`
-	Notification struct {
-		InApp int `json:"inapp"`
-		Email int `json:"email"`
-	} `json:"notification"`
-}
